test(trpc): inline routers per test instead of a shared fixture
Each test now creates only the router it needs; tests/shared.ts keeps
just the tRPC builder and schemas.
