+++ packages/trpc/src/to-orpc-router.test-d.ts
@@ -2,8 +2,61 @@ import type { InferRouterInitialContext, Procedure, Router, Schema } from '@orpc
 import type { AsyncIteratorClass } from '@orpc/shared'
 import type { inferRouterContext } from '@trpc/server'
 import type { TrackedData } from '@trpc/server/unstable-core-do-not-import'
-import type { TRPCContext, trpcRouter } from '../tests/shared'
+import type { TRPCContext } from '../tests/shared'
 import type { ToORPCRouterResult } from './to-orpc-router'
+import { lazy, tracked, TRPCError } from '@trpc/server'
+import * as z from 'zod'
+import { inputSchema, outputSchema, t } from '../tests/shared'
+
+const trpcRouter = t.router({
+  ping: t.procedure
+    .input(inputSchema)
+    .output(outputSchema)
+    .query(({ input }) => {
+      return { output: Number(input.input) }
+    }),
+
+  throw: t.procedure
+    .input(z.object({ b: z.number(), c: z.string() }))
+    .query(() => {
+      throw new TRPCError({
+        code: 'PARSE_ERROR',
+        message: 'throw',
+      })
+    }),
+
+  subscribe: t.procedure
+    .input(z.object({ u: z.string() }))
+    .subscription(async function* () {
+      yield 'pong'
+      yield tracked('id-1', { order: 1 })
+    }),
+
+  nested: {
+    ping: t.procedure
+      .input(z.object({ a: z.string() }))
+      .output(z.string().transform(val => Number(val)))
+      .query(({ input }) => {
+        return `1234${input.a}`
+      }),
+  },
+
+  lazy: lazy(() => Promise.resolve({ default: t.router({
+    subscribe: t.procedure
+      .subscription(async function* () {
+        yield 'pong'
+      }),
+
+    lazy: lazy(() => Promise.resolve({ default: t.router({
+      throw: t.procedure
+        .input(inputSchema)
+        .output(outputSchema)
+        .query(() => {
+          throw new Error('lazy.lazy.throw')
+        }),
+    }) })),
+  }) })),
+})
 
 it('ToORPCRouterResult', () => {
   const orpcRouter = {} as ToORPCRouterResult<
